+++ src/main/java/com/sparta/mvm/service/PostService.java
@@ -3,10 +3,14 @@ package com.sparta.mvm.service;
 import com.sparta.mvm.dto.PostRequestDto;
 import com.sparta.mvm.dto.PostResponseDto;
 import com.sparta.mvm.entity.Post;
+import com.sparta.mvm.entity.User;
 import com.sparta.mvm.exception.CustomException;
 import com.sparta.mvm.exception.ErrorEnum;
 import com.sparta.mvm.repository.PostRepository;
+import com.sparta.mvm.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,10 +22,11 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class PostService {
     private final PostRepository postRepository;
-
+    private final UserRepository userRepository;
+    
     public PostResponseDto findById(long postId) {
         Post post = findPostById(postId);
-        return PostResponseDto.toList("게시글 조회 성공 🎉", 200, post);
+        return PostResponseDto.toDto("게시글 조회 성공 🎉", 200, post);
     }
 
     private Post findPostById(long postId) {
@@ -30,10 +35,13 @@ public class PostService {
 
     @Transactional
     public PostResponseDto save(PostRequestDto request) {
-        Post post = postRepository.save(request.toEntity());
+        Long loggedInUserId = getLoggedInUserId();
+        User user = getUserById(loggedInUserId);
+        Post post = request.toEntity();
+        post.setUser(user);
+        post = postRepository.save(post);
         return PostResponseDto.toDto("게시글 등록 성공 🎉", 200, post);
     }
-
 
     @Transactional
     public List<PostResponseDto> getAll() {
@@ -41,22 +49,45 @@ public class PostService {
         return list
                 .stream()
                 .sorted(Comparator.comparing(Post::getCreatedAt).reversed())
-                .map(post -> PostResponseDto.toList("게시글 조회 성공 🎉", 200, post))
+                .map(post -> PostResponseDto.toDto("게시글 조회 성공 🎉", 200, post))
                 .toList();
     }
 
     @Transactional
     public PostResponseDto update(long postId, PostRequestDto request) {
         Post post = findPostById(postId);
-        post.update(request.getContents());
-        return PostResponseDto.toDto("게시글 수정 성공 🎉", 200, post);
+        Long loggedInUserId = getLoggedInUserId();
+        if (loggedInUserId.equals(post.getUser().getId())) {
+            post.update(request.getContents());
+            return PostResponseDto.toDto("게시글 수정 성공 🎉", 200, post);
+        } else {
+            throw new CustomException(ErrorEnum.BAD_AUTH_PUT);
+        }
     }
-
 
     @Transactional
     public PostResponseDto delete(long postId) {
         Post post = findPostById(postId);
-        postRepository.delete(post);
-        return PostResponseDto.toDeleteResponse("게시글 삭제 성공 🎉", 200);
+        Long loggedInUserId = getLoggedInUserId();
+        if (loggedInUserId.equals(post.getUser().getId())) {
+            postRepository.delete(post);
+            return PostResponseDto.toDeleteResponse("게시글 삭제 성공 🎉", 200);
+        } else {
+            throw new CustomException(ErrorEnum.BAD_AUTH_DELETE);
+        }
+    }
+
+    private User getUserById(Long userId)
+    {
+        return userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorEnum.USER_NOT_FOUND));
+    }
+
+    // 현재 로그인한 사용자의 ID 가져오기
+    private Long getLoggedInUserId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ErrorEnum.USER_NOT_FOUND));
+        return user.getId();
     }
 }
